fixed: fix eros-tempalte issue#133 and issue#78
